init rest service predict test
